Added PWM wave generation code
We need to generate a 56.8 kHz wave signal as the operating frequency for
the IR LED. Instead of wasting hardware on a 555 timer or a oscillator we
can just use the PWM functionality of the HCS12. I will admit that I don't
have a full grasp of what is going on in the PWMSET subroutine. Will have
to ask the professor later since the registers are not in the
documentation.